+++ android/src/main/java/com/mrousavy/camera/skia/SkiaPreviewView.kt
@@ -2,13 +2,11 @@ package com.mrousavy.camera.skia
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.graphics.ImageFormat
 import android.graphics.SurfaceTexture
-import android.opengl.EGL14
-import android.opengl.GLES11Ext
-import android.opengl.GLES20
-import android.opengl.GLSurfaceView
-import android.os.HandlerThread
+import android.hardware.camera2.CameraManager
 import android.util.Log
+import android.util.Size
 import android.view.Choreographer
 import android.view.Surface
 import android.view.SurfaceHolder
@@ -16,22 +14,18 @@ import android.view.SurfaceView
 import com.facebook.jni.HybridData
 import com.facebook.proguard.annotations.DoNotStrip
 import com.mrousavy.camera.CameraQueues
+import com.mrousavy.camera.extensions.getPreviewSize
 import java.io.Closeable
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
-import java.nio.FloatBuffer
-import java.nio.IntBuffer
-import javax.microedition.khronos.egl.EGLConfig
-import javax.microedition.khronos.opengles.GL
-import javax.microedition.khronos.opengles.GL10
 
 @SuppressLint("ViewConstructor")
 @Suppress("KotlinJniMissingFunction")
 class SkiaPreviewView(context: Context,
+                      cameraManager: CameraManager,
+                      cameraId: String,
                       private val onSurfaceChanged: (surface: Surface?) -> Unit) :
   SurfaceView(context), SurfaceHolder.Callback, Closeable {
   companion object {
-    private val TAG = "SkiaPreviewView"
+    private const val TAG = "SkiaPreviewView"
   }
 
   data class InputTexture(val textureId: Int, val surfaceTexture: SurfaceTexture, val surface: Surface)
@@ -41,11 +35,15 @@ class SkiaPreviewView(context: Context,
   private var inputTexture: InputTexture? = null
   private var isAlive = true
   private val thread = CameraQueues.previewQueue.handler
+  private val previewStreamSize: Size
 
   init {
     Log.i(TAG, "Initializing SkiaPreviewView...")
     mHybridData = initHybrid()
     holder.addCallback(this)
+
+    val cameraCharacteristics = cameraManager.getCameraCharacteristics(cameraId)
+    previewStreamSize = cameraCharacteristics.getPreviewSize(ImageFormat.YUV_420_888)
   }
 
   override fun close() {
@@ -83,6 +81,7 @@ class SkiaPreviewView(context: Context,
     // Create Java part (Surface)
     val textureId = createTexture()
     val surfaceTexture = SurfaceTexture(textureId)
+    surfaceTexture.setDefaultBufferSize(previewStreamSize.width, previewStreamSize.height)
     surfaceTexture.setOnFrameAvailableListener { texture -> onCameraFrame(texture) }
     val surface = Surface(surfaceTexture)
     inputTexture = InputTexture(textureId, surfaceTexture, surface)
@@ -90,7 +89,7 @@ class SkiaPreviewView(context: Context,
     onSurfaceChanged(surface)
 
     // Start updating the Preview View (~60 FPS)
-    this.thread.post {
+    thread.post {
       startLooping(Choreographer.getInstance())
     }
   }
@@ -112,52 +111,6 @@ class SkiaPreviewView(context: Context,
     Log.i(TAG, "surfaceChanged($w, $h)")
     onSurfaceResized(w, h)
   }
-
-  /*
-  override fun onSurfaceCreated(gl: GL10, config: EGLConfig?) {
-    destroyInputTexture(gl)
-
-    Log.i(TAG, "onSurfaceCreated(OpenGL)")
-    val textures = IntBuffer.allocate(1)
-    gl.glGenTextures(1, textures)
-
-    val textureId = textures[0]
-    val surfaceTexture = SurfaceTexture(textureId)
-    surfaceTexture.setDefaultBufferSize(1280, 720)
-    surfaceTexture.setOnFrameAvailableListener { s ->
-      // CAMERA - Drawing at 1..240 FPS
-      Log.i(TAG, "onFrameAvailable(..)")
-      // TODO: 1. s.updateTexImage()
-      // TODO: 2. call Skia Frame Processor
-    }
-    val surface = Surface(surfaceTexture)
-    onSurfaceChanged(surface)
-    inputTexture = InputTexture(textureId, surfaceTexture, surface)
-
-    onSurfaceCreated()
-  }
-
-  override fun onDrawFrame(gl: GL10) {
-    // PREVIEW - Drawing at 60 FPS
-    Log.i(TAG, "onDrawFrame(..)")
-    val inputTexture = inputTexture ?: return
-
-    gl.glClearColor(1f, 0f, 0f, 1f)
-    gl.glClear(GLES20.GL_COLOR_BUFFER_BIT)
-
-    inputTexture.surfaceTexture.updateTexImage()
-
-    onDrawFrame(inputTexture.textureId, 1280, 720)
-  }
-
-  private fun destroyInputTexture(gl: GL10? = null) {
-    val inputTexture = inputTexture ?: return
-
-    inputTexture.surface.release()
-    inputTexture.surfaceTexture.release()
-    gl?.glDeleteTextures(1, intArrayOf(inputTexture.textureId), 0)
-  }
-   */
 
   private external fun initHybrid(): HybridData
   private external fun destroy()
